requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==41.0.4
Django==5.1.4
django-allauth==65.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.2
requests==2.30.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
